+++ lib/src/tag/tagger.dart
@@ -81,6 +81,7 @@ import 'package:path/path.dart' as path;
 import 'package:pub_semver/pub_semver.dart';
 
 import '../null_safety.dart';
+import '../pubspec.dart';
 import '../pubspec_io.dart' show pubspecFromDir;
 import '../utils.dart' show dartFilesFromLib;
 import '_common.dart';
@@ -94,6 +95,7 @@ export '_specs.dart' show Runtime;
 /// Calculates the tags for the package residing in a given directory.
 class Tagger {
   final String packageName;
+  final Pubspec _pubspec;
   final AnalysisSession _session;
   final PubspecCache _pubspecCache;
   final bool _isBinaryOnly;
@@ -115,6 +117,7 @@ class Tagger {
 
   Tagger._(
     this.packageName,
+    this._pubspec,
     this._session,
     PubspecCache pubspecCache,
     this._isBinaryOnly,
@@ -171,6 +174,7 @@ class Tagger {
         .toList();
     return Tagger._(
       pubspec.name,
+      pubspec,
       session,
       pubspecCache,
       isBinaryOnly,
@@ -292,6 +296,14 @@ class Tagger {
       explanations
           .add(Explanation('Tag detection failed.', e.message, tag: null));
       return;
+    }
+  }
+
+  /// Adds tags for Flutter plugins.
+  void flutterPluginTags(List<String> tags, List<Explanation> explanations) {
+    const _pluginTag = 'is:plugin';
+    if (_pubspec.hasFlutterPluginKey) {
+      tags.add(_pluginTag);
     }
   }